requirements.txt
streamlit
openai
pytesseract
Pillow
pyzbar